% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dslanguages.R
\docType{data}
\name{dslanguages}
\alias{dslanguages}
\title{Indo-European languages dataset}
\format{
An data frame with 84 rows and 84 columns containing a distance matrix.
}
\usage{
data(dslanguages)
}
\description{
It contains the distances between 84 Indo-European languages based on the mean percent difference in cognacy, using the 200 Swadesh words.
}
\details{
Once the data set is loaded, it can be accessed as an object of class dataframe called \code{dslanguages}.
}
\references{
Dyen, I., Kruskal, J., and Black, P. (1992). An indoeuropean classification: A lexicostatistical experiment. Transactions of the American Philosophical Society. 82, (5).
}
\keyword{datasets}
