% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsyeastexpression.R
\docType{data}
\name{dsyeastexpression}
\alias{dsyeastexpression}
\title{Budding Yeast dataset}
\format{
An data frame with 2467 rows and 79 columns.
}
\source{
\url{https://www.pnas.org/content/suppl/1998/12/08/95.25.14863.DC1/3917data.xls}
}
\usage{
data(dsyeastexpression)
}
\description{
It contains the expression levels of 2467 genes on 79 samples corresponding to 8 different experiments of the budding yeast: alpha factor (18 samples), cdc15 (15 samples), cold shock (4 samples), diauxic shift (7 samples), DTT shock (4 samples), elutriation (14 samples), heat shock (6 samples) and sporulation (11 samples).
}
\details{
Once the data set is loaded, it can be accessed as an object of class dataframe called \code{dsyeastexpression}.
}
\references{
M. B. Eisen, P. T. Spellman, P. O. Brown, and D. Botstein. (1998). Cluster analysis and display of genome-wideexpression patterns.Proceedings of the National Academy of Sciences, 95(25):14863–14868
}
\keyword{datasets}
