% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_results.R
\name{generate.results}
\alias{generate.results}
\title{Generates clustering results}
\usage{
generate.results(g_clusters, distance.matrix)
}
\arguments{
\item{g_clusters}{igraph object with all clusters as connected components.}

\item{distance.matrix}{The original distance matrix.}
}
\value{
A list with cnumber, cluster, partition, csize, network.
}
\description{
Generates clustering results
}
\keyword{internal}
