% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_single_graphs.R
\name{only.single.graphs}
\alias{only.single.graphs}
\title{Generates the solution when only singletons are yielded}
\usage{
only.single.graphs(total_nodos, nodos_singletons)
}
\arguments{
\item{total_nodos}{Total number of nodes in data matrix.}

\item{nodos_singletons}{Nodes list with cluster singletons.}
}
\value{
An object of class "igraph" as a network representing the clustering solution.
}
\description{
Generates the solution when only singletons are yielded
}
\keyword{internal}
